Reworked caching in Spline
A spline now has an accuracy property; for each of spline's methods that require an accuracy parameter (i.e. GetFlattened, GetLength), if no accuracy parameter is passed in the spline's accuracy property is used instead. Only in that case (or if the passed in value is equal to the spline's accuracy) will the result be cached.