Added PosRotScale and implemented it
A spline has a transform consisting of position, rotation and scale. As an invariant, all points in the spline already have this transform applied to them. Because of that, to change i.e. the position of the spline, the current position first is unapplied from each point and then the new one is applied. This decision was made to prevent additional overhead involved in getting a point's position, which should be a way more common operation than changing the spline transform. If someone still wants to get the 'raw' value of a point, they can use SplineUtility.UnapplyPosRotScale.